New popup is work fine
