+++ Tik-Tok Lite/BackgroundCleanerView.swift
@@ -1,0 +1,20 @@
+//
+//  BackgroundCleanerView.swift
+//  Tik-Tok Lite
+//
+//  Created by Fantom on 16.01.2022.
+//
+
+import SwiftUI
+
+struct BackgroundCleanerView: UIViewRepresentable {
+    func makeUIView(context: Context) -> UIView {
+        let view = UIView()
+        DispatchQueue.main.async {
+            view.superview?.superview?.backgroundColor = .clear
+        }
+        return view
+    }
+
+    func updateUIView(_ uiView: UIView, context: Context) {}
+}
